Bump WP Mock's phpunit version requirements
The library already requires version 4.3. There are now a few features of 4.3+ that we have needed in tests, and now in the library itself. Deprecating wpFunction proved to be far too difficult if we were to maintain backwards compatibility with 3.7, so we're updating the requirements to 4.3. We could have done 4.0, but most people using a version < 4.3 are going to be using 3.7 anyway. It's easier to consolidate the version requirements so as to avoid confusion and to decrease the likelihood that our tests will miss an integration issue by not being able to run on the lowest supported version of phpunit.