Add GPLv2 License Text
The WP_Mock project is licensed under the terms of the GNU General Public License, version 2 (or, at your discretion, any later version). We're adding the text of the GPL to the repository to avoid any possible confusion with how or why the project is licensed as it is.

Props @jhabdas, @johnpbloch, @emeaguiar