Add gitbook.yaml (#201)
<!--
Filling out the required portions of this template is mandatory. 
Any PR that does not include enough information to be reviewed may be
closed at a maintainers' discretion.
All new code requires associated documentation and unit tests.
-->

# Summary <!-- Required -->

Adds a `gitbook.yaml` file to exclude syncing doc files outside of the
`docs` folder in the git repository root.

Will also point to an `intrroduction.md` file as the main docs page (to
avoid confusion with readme.md).

Readme, changelog etc. in the git root are not expected to be handled by
GitBook.

See #199.

### Reviewer checklist <!-- Required -->

<!-- The following checklist is for the reviewer: add any steps that may
be relevant while reviewing this pull request -->

- [ ] Code changes review
- [x] Documentation changes review
- [x] Unit tests pass
- [x] Static analysis passes